builder_mcp/gateway/requirements.txt
# MCP Gateway Service
# Install into your conda environment:
#   conda activate aihub2
#   pip install -r builder_mcp\gateway\requirements.txt

fastapi>=0.100.0          # Web framework (async-native)
uvicorn>=0.23.0           # ASGI server for FastAPI
httpx>=0.24.0             # Async HTTP client for remote SSE servers
pydantic>=2.0.0           # Data validation (comes with FastAPI)
python-dotenv             # Environment variable loading
requests                  # Sync HTTP (for any outbound calls)
